fix: Suppress restriction lints in test/bench code and fix standard clippy warnings
Add #[expect(...)] attributes to test modules and integration test files
for restriction lints (unwrap_used, indexing_slicing, expect_used, panic,
todo, unimplemented, unreachable) that are intentional in test code. Uses
#[expect] over #[allow] so unfulfilled suppressions produce warnings.

Fix standard clippy lints directly: redundant closures replaced with
function references, &Vec<T> changed to &[T], clone on Copy types
changed to dereference, unnecessary cast removed, unused lifetimes
elided, vec! replaced with array literal where Vec not needed.

Suppress intentional benchmark patterns (identity_op, bind_instead_of_map,
unnecessary_map_or, unnecessary_fold) with #[expect] on module
declarations since these mirror fp library operations for fair comparison.

Add missing re-exports of bimap, bi_fold_left, bi_fold_right,
bi_fold_map, bi_traverse from functions.rs, fixing 5 failing doc tests.

+++ fp-library/benches/benchmarks/lazy.rs
@@ -328,7 +328,7 @@ pub fn bench_lazy(c: &mut Criterion) {
 								.bind(move |x| build_right(x - 1).bind(|y| Free::pure(y + 1)))
 						}
 					}
-					build_right(d as i64).evaluate()
+					build_right(d).evaluate()
 				})
 			});
 		}
